Profile GUI

+++ body/controllers/batches.php
@@ -28,12 +28,22 @@ class batches extends CI_Controller {
             $batch = new Batch();
 
             if ($_FILES['batch_image']['name'] != '') {
-                $image = $this->uploadImage();
+                $image = $this->uploadImage('batch_image');
                 if (isset($image['error'])) {
                     $this->session->set_flashdata('file_errors', $image['error']);
                     redirect(base_url() . 'batch/add', 'refresh');
                 } else if (isset($image['upload_data'])) {
                     $batch->image = $image['upload_data']['file_name'];
+                }
+            }
+
+            if ($_FILES['batch_cover_image']['name'] != '') {
+                $image = $this->uploadImage('batch_cover_image');
+                if (isset($image['error'])) {
+                    $this->session->set_flashdata('file_errors_cover', $image['error']);
+                    redirect(base_url() . 'batch/add', 'refresh');
+                } else if (isset($image['upload_data'])) {
+                    $batch->cover_image = $image['upload_data']['file_name'];
                 }
             }
 
@@ -65,17 +75,30 @@ class batches extends CI_Controller {
                 $batch->where('id', $id)->get();
 
                 if ($_FILES['batch_image']['name'] != '') {
-                    $image = $this->uploadImage();
+                    $image = $this->uploadImage('batch_image');
                     if (isset($image['error'])) {
                         $this->session->set_flashdata('file_errors', $image['error']);
                         redirect(base_url() . 'batch/edit/' . $id, 'refresh');
                     } else if (isset($image['upload_data'])) {
 
-                        if (file_exists('assets/img/batches/' . $batch->image)) {
+                        if (!is_null($batch->image) && file_exists('assets/img/batches/' . $batch->image)) {
                             unlink('assets/img/batches/' . $batch->image);
                         }
 
                         $batch->image = $image['upload_data']['file_name'];
+                    }
+                }
+
+                if ($_FILES['batch_cover_image']['name'] != '') {
+                    $image = $this->uploadImage('batch_cover_image');
+                    if (isset($image['error'])) {
+                        $this->session->set_flashdata('file_errors_cover', $image['error']);
+                        redirect(base_url() . 'batch/edit/' . $id, 'refresh');
+                    } else if (isset($image['upload_data'])) {
+                        if (!is_null($batch->cover_image) && file_exists('assets/img/batches/cover_image/' . $batch->cover_image)) {
+                            unlink('assets/img/batches/cover_image/' . $batch->cover_image);
+                        }
+                        $batch->cover_image = $image['upload_data']['file_name'];
                     }
                 }
 
@@ -113,6 +136,9 @@ class batches extends CI_Controller {
             if (file_exists('assets/img/batches/' . $batch->image)) {
                 unlink('assets/img/batches/' . $batch->image);
             }
+            if (!is_null($batch->cover_image) && file_exists('assets/img/batches/cover_image/' . $batch->cover_image)) {
+                unlink('assets/img/batches/cover_image/' . $batch->cover_image);
+            }
             $batch->delete();
             $this->session->set_flashdata('success', $this->lang->line('delete_data_success'));
             redirect(base_url() . 'batch', 'refresh');
@@ -122,19 +148,41 @@ class batches extends CI_Controller {
         }
     }
 
-    function uploadImage() {
-        $this->upload->initialize(array(
-            'upload_path'   => "./assets/img/batches/",
-            'allowed_types' => 'jpg|jpeg|gif|png|bmp',
-            'overwrite' => FALSE,
-            'remove_spaces' => TRUE,
-            'encrypt_name' => TRUE
-        ));
+    function uploadImage($field) {
+        if($field == 'batch_image'){
+            $this->upload->initialize(array(
+                'upload_path'   => "./assets/img/batches/",
+                'allowed_types' => 'jpg|jpeg|gif|png|bmp',
+                'overwrite' => FALSE,
+                'remove_spaces' => TRUE,
+                'encrypt_name' => TRUE
+            ));
+        }
 
-        if (!$this->upload->do_upload('batch_image')) {
+        if($field == 'batch_cover_image'){
+            $this->upload->initialize(array(
+                'upload_path'   => "./assets/img/batches/cover_image/",
+                'allowed_types' => 'jpg|jpeg|gif|png|bmp',
+                'overwrite' => FALSE,
+                'remove_spaces' => TRUE,
+                'encrypt_name' => TRUE
+            ));
+        }
+        
+
+        if (!$this->upload->do_upload($field)) {
             $data = array('error' => $this->upload->display_errors());
         } else {
-            $data = array('upload_data' => $this->upload->data('batch_image'));
+            $data = array('upload_data' => $this->upload->data($field));
+
+            if ($field == 'batch_cover_image' && $data['upload_data']['file_name'] != '') {
+                $image = str_replace(' ', '_', $data['upload_data']['file_name']);
+                $this->load->helper('image_manipulation/image_manipulation');
+                include_lib_image_manipulation();
+                $magicianObj = new imageLib('./assets/img/batches/cover_image/' . $image);
+                $magicianObj->resizeImage(675, 280, 'exact');
+                $magicianObj->saveImage('./assets/img/batches/cover_image/' . $image, 100);
+            }
         }
 
         return $data;
